requirements.txt
numpy
opencv-python
torch
python-telegram-bot
onnxruntime
torchvision